Fix: Traffic light component and UI layout
